+++ src/features/autocomplete.js
@@ -1,12 +1,6 @@
 import SessionManager from '../core/session-manager.js';
 import ErrorHandler from '../utils/error-handler.js';
-import {APPS_MENU_CONFIG} from './menuItems/apps.menu.js';
-import { ESD_MENU_CONFIG } from './menuItems/esd.menu.js';
-import {FLOWS_MENU_CONFIG} from './menuItems/flows.menu.js';
-import { LABELS_MENU_CONFIG } from './menuItems/labels.menu.js';
-import {OBJECTS_MENU_CONFIG} from './menuItems/objects.menu.js';
-import {PERMSETS_MENU_CONFIG} from './menuItems/permsets.menu.js';
-import {PROFILES_MENU_CONFIG} from './menuItems/profiles.menu.js';
+import {MENU_CONFIGS} from './menuItems/index.js';
 import NavigatorService from './navigator.service.js';
 
 // Define constants for better maintainability
@@ -19,6 +13,8 @@ const STATES = {
 	LABEL_SELECTED: 'LABELS_SELECTED',
 	PERMSET_SELECTED: 'PERMSETS_SELECTED',
 	ESD_SELECTED: 'ESD_SELECTED',
+	CUSTOMMETADATA_SELECTED: 'CUSTOMMETADATA_SELECTED',
+	CUSTOMSETTING_SELECTED: 'CUSTOMSETTINGS_SELECTED',
 	BOTS_SELECTED: 'BOTS_SELECTED'
 };
 
@@ -32,18 +28,12 @@ const STATE_STORE = {
 	[STATES.LABEL_SELECTED]: {CONFIG: {}, DATA: []},
 	[STATES.PERMSET_SELECTED]: {CONFIG: {}, DATA: []},
 	[STATES.ESD_SELECTED]: {CONFIG: {}, DATA: []},
+	[STATES.CUSTOMMETADATA_SELECTED]: {CONFIG: {}, DATA: []},
+	[STATES.CUSTOMSETTING_SELECTED]: {CONFIG: {}, DATA: []},
 	[STATES.BOTS_SELECTED]: {CONFIG: {}, DATA: []},
 };
 
-const newEntityHandlers = {
-	objects: OBJECTS_MENU_CONFIG,
-	apps: APPS_MENU_CONFIG,
-	profiles: PROFILES_MENU_CONFIG,
-	flows: FLOWS_MENU_CONFIG,
-	permsets: PERMSETS_MENU_CONFIG,
-	labels: LABELS_MENU_CONFIG,
-	esd: ESD_MENU_CONFIG
-};
+const newEntityHandlers = MENU_CONFIGS;
 
 class AutocompleteManager {
 	constructor(inputElement, dropdownElement) {